fix bug

+++ feature_select_corr_ML/shared.py
@@ -8,7 +8,7 @@ import matplotlib.pyplot as plt
 
 
 def prepare_data(num):
-    features = pd.read_csv('../../elliptic_bitcoin_dataset/full_data.csv',header=None, dtype='float64')
+    features = pd.read_csv('../../elliptic_bitcoin_dataset/full_data.csv',header=None)
     classes = pd.read_csv('../../elliptic_bitcoin_dataset/elliptic_txs_classes.csv')
     feature = [str(i) for i in range(171)]
     features.columns = ["txId","time_step"] + feature
