+++ raw_ML/shared.py
@@ -7,7 +7,7 @@ from sklearn.model_selection import train_test_split
 
 
 def prepare_data():
-    features = pd.read_csv('../../elliptic_bitcoin_dataset/elliptic_txs_features.csv',header=None, dtype='float64')
+    features = pd.read_csv('../../elliptic_bitcoin_dataset/elliptic_txs_features.csv',header=None)
     classes = pd.read_csv('../../elliptic_bitcoin_dataset/elliptic_txs_classes.csv')
     #edges = pd.read_csv('../../elliptic_bitcoin_dataset/elliptic_txs_edgelist.csv')
     feature = [str(i) for i in range(165)]
