+++ stat/hy_testing_raw.py
@@ -7,7 +7,7 @@ import statsmodels.api as sm
 from scipy import stats
 import matplotlib.pyplot as plt
 import seaborn as sns
-features = pd.read_csv('../../elliptic_bitcoin_dataset/elliptic_txs_features.csv',header=None, dtype='float64')
+features = pd.read_csv('../../elliptic_bitcoin_dataset/elliptic_txs_features.csv',header=None)
 classes = pd.read_csv('../../elliptic_bitcoin_dataset/elliptic_txs_classes.csv')
 feature = [str(i) for i in range(165)]
 features.columns = ["txId","time_step"] + feature
